+++ risc0/zkvm/src/host/api/tests.rs
@@ -49,10 +49,6 @@ impl TestClientConnector {
 }
 
 impl Connector for TestClientConnector {
-    fn get_version(&self) -> Result<semver::Version> {
-        unimplemented!()
-    }
-
     fn connect(&self) -> Result<ConnectionWrapper> {
         let (stream, _) = self.listener.accept()?;
         Ok(ConnectionWrapper::new(Box::new(TcpConnection::new(stream))))
